refactor : 매직 넘버 사용 및 최신순 정렬
