+++ src/main/java/com/programmers/bucketback/domains/inventory/application/InventoryReader.java
@@ -1,5 +1,6 @@
 package com.programmers.bucketback.domains.inventory.application;
 
+import java.util.Comparator;
 import java.util.List;
 
 import org.springframework.stereotype.Component;
@@ -9,13 +10,13 @@ import com.programmers.bucketback.domains.common.Hobby;
 import com.programmers.bucketback.domains.common.MemberUtils;
 import com.programmers.bucketback.domains.common.vo.CursorPageParameters;
 import com.programmers.bucketback.domains.inventory.api.dto.response.InventoryInfoSummary;
+import com.programmers.bucketback.domains.inventory.application.vo.InventoryProfile;
 import com.programmers.bucketback.domains.inventory.domain.Inventory;
 import com.programmers.bucketback.domains.inventory.domain.InventoryItem;
 import com.programmers.bucketback.domains.inventory.repository.InventoryItemRepository;
 import com.programmers.bucketback.domains.inventory.repository.InventoryRepository;
 import com.programmers.bucketback.domains.item.application.ItemReader;
 import com.programmers.bucketback.domains.item.domain.Item;
-import com.programmers.bucketback.domains.member.application.InventoryProfile;
 import com.programmers.bucketback.domains.review.application.ReviewReader;
 import com.programmers.bucketback.domains.review.domain.Review;
 import com.programmers.bucketback.global.error.exception.EntityNotFoundException;
@@ -30,6 +31,8 @@ import lombok.extern.slf4j.Slf4j;
 @Transactional(readOnly = true)
 public class InventoryReader {
 
+	private static final int ITEM_IMAGE_LIMIT = 4;
+	private static final int INVENTORY_PROFILE_LIMIT = 3;
 	private final InventoryRepository inventoryRepository;
 	private final InventoryItemRepository inventoryItemRepository;
 	private final ReviewReader reviewReader;
@@ -119,11 +122,14 @@ public class InventoryReader {
 
 	private List<InventoryProfile> selectInventoryProfile(final List<Inventory> inventories) {
 		List<Inventory> selectedInventories = inventories.stream()
-			.limit(3)
+			.sorted(Comparator.comparing(Inventory::getModifiedAt).reversed())
+			.limit(INVENTORY_PROFILE_LIMIT)
 			.toList();
+
 		return selectedInventories.stream()
 			.map(inventory -> {
 				List<String> itemImages = extractInventoryItemImages(inventory);
+
 				return InventoryProfile.of(inventory, itemImages);
 			})
 			.toList();
@@ -131,7 +137,8 @@ public class InventoryReader {
 
 	private List<String> extractInventoryItemImages(final Inventory inventory) {
 		return inventory.getInventoryItems().stream()
-			.limit(4)
+			.sorted(Comparator.comparing(InventoryItem::getModifiedAt).reversed())
+			.limit(ITEM_IMAGE_LIMIT)
 			.map(InventoryItem::getItem)
 			.map(Item::getImage)
 			.toList();
